chore(arena): move event related types and utilities to its own file
